pyproject.toml
[project]
name = "trudy-telegram"
version = "0.1.0"
description = "Telegram bot-based note-taking system with AI processing"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Abhinav Jain" }
]

dependencies = [
    # Telegram
    "python-telegram-bot>=21.0",
    "aiohttp>=3.9.0",

    # Media processing
    "ffmpeg-python>=0.2.0",

    # Article extraction
    "trafilatura>=1.8.0",
    "newspaper3k>=0.2.8",
    "beautifulsoup4>=4.12.0",

    # YouTube
    "youtube-transcript-api>=0.6.0",
    "pytube>=15.0.0",

    # AI/ML
    "ollama>=0.1.0",

    # Configuration & utilities
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # CLI
    "click>=8.1.0",
    "rich>=13.0.0",

    # Utils
    "python-dateutil>=2.8.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
trudy = "src.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
